Recent advances in deep learning based sentiment analysis
Sentiment analysis is one of the most popular research areas in natural language processing. It is extremely useful in many applications, such as social media monitoring and e-commerce. Recent application of deep learning based methods has dramatically changed the research strategies and improved the performance of many traditional sentiment analysis tasks, such as sentiment classification and aspect based sentiment analysis. Moreover, it also pushed the boundary of various sentiment analysis task, including sentiment classification of different text granularities and in different application scenarios, implicit sentiment analysis, multimodal sentiment analysis and generation of sentiment-bearing text. In this paper, we give a brief introduction to the recent advance of the deep learning-based methods in these sentiment analysis tasks, including summarizing the approaches and analyzing the dataset. This survey can be well suited for the researchers studying in this field as well as the researchers entering the field.